Create QUALIFICATIONS INSTRUCTION
NOTES ON USE
This instruction should be given in addition to and at the conclusion of the instructions normally given to the prospective jurors.  The portion of this instruction dealing with communication with others and outside research may need to be modified to include other specified means of communication or research as technology develops.
This instruction was adopted in 2010 [52 So. 3d 595] and amended in 2014.