Create 2.1 Preliminary Instructions
Comments

Florida Rule of Judicial Administration 2.451 directs trial judges to instruct jurors on the use of cell phones and other electronic devices. During the trial, the trial judge may remove the jurors’ cell phones or other electronic devices.  The trial judge also has the option to allow the jurors to keep the cell phones and electronic devices during trial until the jurors begin deliberations. Rule 2.451 prohibits jurors from using the cell phones or electronic devices to find out information about the case or to communicate with others about the case. The jurors also cannot use the electronic devices to record, photograph, or videotape the proceedings. In recognition of the discretion rule 2.451 gives trial judges, this instruction provides two alternatives:  (A) requiring jurors to turn off electronic devices during court proceedings and removing their cell phones and electronic devices during deliberations; or (B) removing the cell phones and electronic devices during all proceedings and deliberations. These instructions may be modified to fit the practices of a trial judge in a particular courtroom. These instructions are not intended to limit the discretion of the trial court to control the proceedings.  
The portion of this instruction dealing with communication with others and outside research may need to be modified to include other specified means of communication or research as technology develops.
This instruction was adopted in 1981 and amended in 2010 [52 So. 3d 595] and 2014.